+++ config/runtime.exs
@@ -8,6 +8,10 @@ config :td_dd,
 
 config :td_cluster, groups: [:dd]
 
+config :td_cache, :audit, maxlen: System.get_env("REDIS_AUDIT_STREAM_MAXLEN", "100")
+
+config :td_cache, :event_stream, maxlen: System.get_env("REDIS_STREAM_MAXLEN", "100")
+
 if config_env() == :prod do
   get_ssl_option = fn env_var, option_key ->
     if System.get_env("DB_SSL", "") |> String.downcase() == "true" do