+++ lib/td_dd/data_structures/data_structures.ex
@@ -707,6 +707,13 @@ defmodule TdDd.DataStructures do
     end
   end
 
+  def get_version_count!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :versions) do
+      %{data_structure: %{versions: versions}} ->
+        length(versions)
+    end
+  end
+
   def get_grants(%DataStructureVersion{data_structure_id: id, path: path}, clauses \\ %{}) do
     ids = Enum.reduce(path, [id], fn %{"data_structure_id" => id}, acc -> [id | acc] end)
 
@@ -723,6 +730,18 @@ defmodule TdDd.DataStructures do
       end
     end)
     |> Enum.sort_by(& &1.id)
+  end
+
+  def get_grants_count(%DataStructureVersion{data_structure_id: id, path: path}) do
+    ids = Enum.reduce(path, [id], fn %{"data_structure_id" => id}, acc -> [id | acc] end)
+
+    dsv_preloader = &enriched_structure_versions(data_structure_ids: &1)
+
+    %{}
+    |> Map.put(:data_structure_ids, ids)
+    |> Map.put(:preload, [:system, :data_structure, data_structure_version: dsv_preloader])
+    |> Grants.list_active_grants()
+    |> length()
   end
 
   def get_grant(%DataStructureVersion{} = dsv, user_id) do
@@ -816,6 +835,20 @@ defmodule TdDd.DataStructures do
     |> where([i], is_nil(i.deleted_at))
     |> where([i], i.data_structure_id == ^id)
     |> select([i], count(i.implementation_id, :distinct))
+    |> Repo.one!()
+  end
+
+  def get_parents_count(%DataStructureVersion{id: id}) do
+    DataStructureRelation
+    |> where([r], r.child_id == ^id)
+    |> select([r], count(r.parent_id, :distinct))
+    |> Repo.one!()
+  end
+
+  def get_children_count(%DataStructureVersion{id: id}) do
+    DataStructureRelation
+    |> where([r], r.parent_id == ^id)
+    |> select([r], count(r.child_id, :distinct))
     |> Repo.one!()
   end
 
@@ -954,7 +987,7 @@ defmodule TdDd.DataStructures do
       ) do
     now = DateTime.utc_now()
 
-    ds_changeset = DataStructure.changeset_updated_at(ds, user_id)
+    ds_changeset = DataStructure.changeset_last_change_by(ds, user_id)
 
     Multi.new()
     |> Multi.run(:descendents, fn _, _ -> get_structure_descendents(data_structure_version) end)
@@ -968,7 +1001,7 @@ defmodule TdDd.DataStructures do
       fn changes -> delete_metadata_descendents(changes) end,
       set: [deleted_at: now, updated_at: now]
     )
-    |> Multi.update(:update_at_change, ds_changeset)
+    |> Multi.update(:last_change, ds_changeset)
     |> Multi.run(:audit, Audit, :data_structure_deleted, [user_id])
     |> Repo.transaction()
     |> on_delete()
@@ -1364,6 +1397,16 @@ defmodule TdDd.DataStructures do
   end
 
   def enriched_structure_version(other, _opts), do: other
+
+  def update_last_change_at(["data_structure:" <> ds_id]) do
+    [ids: [ds_id]]
+    |> list_data_structures()
+    |> List.first()
+    |> DataStructure.changeset_last_change_at(DateTime.utc_now())
+    |> Repo.update()
+  end
+
+  def update_last_change_at(ds_id), do: {:ok, ds_id}
 
   def embeddings([]), do: {:ok, []}